Split cli from Datadog library to make it easier to maintain and isolate them within its own domain
